fix(cron): keep regression suites within lint boundaries
